requirements.txt
djangorestframework
markdown
django-filter
django
django-environ
psycopg2-binary
djangorestframework-xml